+++ jsondaora/serializers.py
@@ -16,7 +16,10 @@ def dataclass_asjson(instance: Any) -> bytes:
 def typed_dict_asjson(
     typed_dict: Any, typed_dict_type: _TypedDictMeta
 ) -> bytes:
-    return orjson.dumps(_choose_typed_dict_fields(typed_dict, typed_dict_type))
+    return orjson.dumps(
+        _choose_typed_dict_fields(typed_dict, typed_dict_type),
+        default=OrjsonDefaultTypes.default_function()
+    )
 
 
 def _choose_typed_dict_fields(