+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Fix typed dict serialization
+
 ### 0.11.0 - 2020-04-04
 
  - Improve bool deserialization